// src/generated-api/models/sendApprovalWeekModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type sendApprovalWeekModel = {
    /**
     * Start date of the week to send for approval
     */
    weekStart?: string;
    /**
     * Assigned approval manager id
     */
    assignedManagerId?: string;
};